Slayer: Fix infobox expiry ignored until first combat
If the Slayer plugin's Task InfoBox is enabled and TaskInfoBox Expiry is set, the timer for the expiration is not started until killing the first slayer monster. This means that if you are doing something other than slayer after a login you have to uncheck TaskInfoBox to get rid of it.

This PR makes it so if a timeout is set, it will always be adhered to.