profiles: add checking for excess colons in notes, usernames and passwords
This handles colons in passwords appropriately by limiting the split length to 3, it will be 2 if there is no password, and 3 if there is a password, however if the password contained a colon it would split further destroying the password contents into separate array elements. This new code keeps the full password in one array element.